bump(api_query): added authenticated user type
