+++ api_query/src/index.ts
@@ -12,10 +12,36 @@ import { UserQuery } from "./user";
 import { UserRegionQuery } from "./user_region";
 import { QueryError } from "./query_error";
 import { Result, Parameter, Element } from "./query";
+import { Prisma } from "@selab-2/groep-1-orm";
+
+type AuthenticatedUser = Prisma.UserGetPayload<{
+    select: {
+        id: true;
+        email: true;
+        first_name: true;
+        last_name: true;
+        last_login: true;
+        date_added: true;
+        phone: true;
+        address_id: true;
+        student: true;
+        super_student: true;
+        admin: true;
+        hash: false;
+        salt: false;
+        address: true;
+        syndicus: {
+            select: {
+                id: true;
+            };
+        };
+    };
+}>;
 
 export {
     ActionQuery,
     AddressQuery,
+    AuthenticatedUser,
     BuildingQuery,
     GarbageQuery,
     ProgressQuery,
