+++ web/package.json
@@ -15,7 +15,7 @@
     "@bitpatty/imgproxy-url-builder": "^1.3.3",
     "@mdi/font": "7.0.96",
     "@selab-2/groep-1-orm": "^1.0.12",
-    "@selab-2/groep-1-query": "^1.0.14",
+    "@selab-2/groep-1-query": "^1.0.15",
     "@vuepic/vue-datepicker": "^4.2.0",
     "chance": "^1.1.11",
     "pinia": "^2.0.35",
